完善了权限管理

+++ src/dao/admin/AccountDao.java
@@ -55,7 +55,7 @@ public class AccountDao {
         return  DbUtil.delete(conn,"account",conditions);
     }
 
-    public static DaoUpdateResult permit(Connection conn, long id,byte permission) {
+    public static DaoUpdateResult permit(Connection conn, long id,int permission) {
         String sql = "update account set permission=? where id=?";
         Object []params = {permission,id};
         //调用DbUtil封装的update方法
